Animations, layout and components
A new layout for the navbar was implemented were there's no main page, now the home page is the main page and the link for it was removed. Animation was added for when you switch from one link to another one (this feature still is been worked on) and the Portfolio and Contact components were added in order to test the previous animations.