SCSS, HammerJS and Contact
Migration to SCSS was done since I consider, it would be better for the app. All SASS file were renamed to be SCSS and syntax was fixed for each of the files. Additional was added HammerJS for more advance animations with gestures and ContactComponent was styled a tittle bit (Still working on it).