Skill bar added
A template image was added in the about section, this image it will be a photo of me. Below a skill bar was added and some transitions. However, this still is being worked on since I have to update the timing when it scrolls down and think of a way in order to make the perecentage dynamic.