Projects added
The project component was added. this will show each component once is clic in the portfolio. Now some test projects are shown in there and a JSON was added with some testing projects (This JSON will be an HTTP Client in the future). added as well some services to get the data and a model of how the project will be. Still being worked.