Project and cards
In order to standardize all card a component Card for the project was creates and added to the portfolio and to home, in home was fixed issues if there was not enough projects, the animation was updated and now has a fade effect at the end. Add card is completely working, just missing a few UI things lastly in the footer when click in let's work together goes back up.